ci: use version 0.0.0 for alpha publish and tag as alpha (#2)
# Description

This PR updates the `alpha-publish` workflow. When an alpha version is published, the base version will be 0.0.0 with the tag of `alpha`.

This would indicate more explicitly that the alpha versions should not be used unless they're in a development or testing environment. The tag `alpha` will not interfere with the `latest` tag (the default tag) that the users install by default when no version is given.

# Checklist

<!-- Append "N/A" at the end of the item to indicate if it's not applicable -->
<!-- E.g. - [ ] ~updated storybook~ N/A, no new components -->
<!-- Prepend "POST-MERGE: " to indicate it will be completed after the merge -->
<!-- E.g. - [ ] POST-MERGE: ramp the feature flag -->

- [x] My PR represents one logical piece of work
- [x] I have performed a self-review of my own code
- [ ] I have made corresponding changes to the documentation N/A
- [ ] I have added tests that prove my fix is effective or that my feature works N/A