feat: remove /api/v2 from the trackingApiBaseUrl (#4)
# Description

Remove `/api/v2` from the `TrackerOptions.trackingApiBaseUrl`. This would give us more flexibility in the future if we use the endpoints not under `/api/v2` (such as other versions of the API).

# Checklist

<!-- Append "N/A" at the end of the item to indicate if it's not applicable -->
<!-- E.g. - [ ] ~updated storybook~ N/A, no new components -->
<!-- Prepend "POST-MERGE: " to indicate it will be completed after the merge -->
<!-- E.g. - [ ] POST-MERGE: ramp the feature flag -->

- [x] My PR represents one logical piece of work
- [x] I have performed a self-review of my own code
- [x] I have made corresponding changes to the documentation
- [x] I have added tests that prove my fix is effective or that my feature works

BREAKING CHANGE: the default `TrackerOptions.trackingApiBaseUrl` no longer has `/api/v2` appended. When a request is sent by the `Tracker` class, it will append `/api/v2` to the `trackingApiBaseUrl`.