fix: identified user data timezone should be timezone offset; add ianaTimezone as a new reserved property (#5)
# Description

The `timezone` property in the reserved user data should be a number representing the timezone offset (e.g. 11 for AEDT) but instead, an IANA timezone identifier is given. This will break existing users' CDPs.

This PR changes that property and is set to the timezone offset. It also now adds a new reserved property `ianaTimezone` which returns the IANA timezone identifier (e.g. Australia/Sydney).

# Checklist

<!-- Append "N/A" at the end of the item to indicate if it's not applicable -->
<!-- E.g. - [ ] ~updated storybook~ N/A, no new components -->
<!-- Prepend "POST-MERGE: " to indicate it will be completed after the merge -->
<!-- E.g. - [ ] POST-MERGE: ramp the feature flag -->

- [x] My PR represents one logical piece of work
- [x] I have performed a self-review of my own code
- [x] I have made corresponding changes to the documentation
- [x] I have added tests that prove my fix is effective or that my feature works